requirements.txt
greenlet==1.1.2
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
